% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_barlow.r, R/theme_inter.r,
%   R/theme_lato.r, R/theme_nunito.r, R/theme_plex.r
\name{import_barlow}
\alias{import_barlow}
\alias{import_inter}
\alias{import_lato}
\alias{import_nunito}
\alias{import_plex}
\title{Import font for use in ggplot2 graphs}
\usage{
import_barlow()

import_inter()

import_lato()

import_nunito()

import_plex()
}
\description{
Import font for use in ggplot2 graphs
}
