% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.r
\name{list_fonts}
\alias{list_fonts}
\title{List all fonts}
\usage{
list_fonts()
}
\value{
Returns a data.frame with name, family, font, weight and path. The name is used as 'family' when plotting.
}
\description{
Lists all fonts and families.
}
