% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_morris.r
\name{palette_morris}
\alias{palette_morris}
\title{10 color qualitative color palette from morris}
\usage{
palette_morris()
}
\description{
10 color qualitative color palette from morris
}
\examples{
library(scales)
scales::show_col(palette_morris()(10))

}
