% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_okabeito.r
\name{palette_okabeito}
\alias{palette_okabeito}
\title{8 color qualitative color palette from okabeito}
\usage{
palette_okabeito()
}
\description{
8 color qualitative color palette from okabeito
}
\examples{
library(scales)
scales::show_col(palette_okabeito()(8))

}
