% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_strong.r
\name{palette_strong}
\alias{palette_strong}
\title{17 color qualitative color palette from strong}
\usage{
palette_strong()
}
\description{
17 color qualitative color palette from strong
}
\examples{
library(scales)
scales::show_col(palette_strong()(17))

}
