% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.r
\name{palette_tableau}
\alias{palette_tableau}
\title{20 color qualitative color palette from tableau}
\usage{
palette_tableau()
}
\description{
20 color qualitative color palette from tableau
}
\examples{
library(scales)
scales::show_col(palette_tableau()(20))

}
