% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plummy.R
\name{plummy}
\alias{plummy}
\title{Custom ggplot2 theme}
\usage{
plummy(style = "grid", family = "", ...)
}
\arguments{
\item{style}{A theme style. Options are 'grid'.}

\item{family}{A font family name. Options are 'barlow', 'inter', 'lato', 'nunito' and 'plex'.}

\item{...}{Arguments passed to low level theme functions theme_style*_font*()}
}
\description{
Custom ggplot2 theme
}
