% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_custom.r
\name{theme_custom}
\alias{theme_custom}
\title{ggplot2 custom theme using default font.}
\usage{
theme_custom(
  base_family = "",
  base_size = 10,
  scaling = 1.2,
  plot_title_family = "",
  plot_title_size = base_size * scaling^2,
  plot_title_face = "bold",
  plot_title_margin = 8,
  subtitle_family = "",
  subtitle_size = base_size * scaling,
  subtitle_face = "plain",
  subtitle_margin = 13,
  strip_text_family = "",
  strip_text_size = base_size,
  strip_text_face = "bold",
  caption_family = "",
  caption_size = base_size/scaling,
  caption_face = "plain",
  caption_margin = 10,
  axis_text_family = "",
  axis_text_face = "plain",
  axis_text_size = base_size/scaling,
  axis_title_family = base_family,
  axis_title_size = base_size/scaling,
  axis_title_face = "plain",
  axis_title_just = "rt",
  plot_margin = margin(30, 30, 30, 30),
  grid_col = "#cccccc",
  grid = TRUE,
  axis_col = "#cccccc",
  axis = TRUE,
  ticks = FALSE,
  tsc_just = 0,
  leg = "v",
  leg_just = "center"
)
}
\arguments{
\item{base_family, base_size}{Base: font family and size}

\item{scaling}{Font size scaling factor}

\item{plot_title_family, plot_title_face, plot_title_size, plot_title_margin}{Plot title: font family, face, size and margin}

\item{subtitle_family, subtitle_face, subtitle_size}{Plot subtitle: font family, face and size}

\item{subtitle_margin}{Plot subtitle margin bottom (single numeric value)}

\item{strip_text_family, strip_text_face, strip_text_size}{Facet label: font family, face and size}

\item{caption_family, caption_face, caption_size, caption_margin}{Plot caption: font family, face, size and margin}

\item{axis_text_family, axis_text_face}{Axis text: font family and face}

\item{axis_text_size}{Font size of axis text}

\item{axis_title_family, axis_title_face, axis_title_size}{Axis title: font family, face and size}

\item{axis_title_just}{Axis title font justification, one of \verb{[blmcrt]} denoting bottom, left, middle, center, right and top.}

\item{plot_margin}{Plot margin (specify with \code{ggplot2::margin()})}

\item{grid_col}{Grid color}

\item{grid}{Panel grid (\code{TRUE}, \code{FALSE}, or a combination of \code{X}, \code{x}, \code{Y}, \code{y} denoting major and minor gridlines)}

\item{axis_col}{Axis color}

\item{axis}{Add x or y axes? \code{TRUE}, \code{FALSE}, "\code{xy}". Includes axis title and axis text.}

\item{ticks}{Adds ticks if \code{TRUE}}

\item{tsc_just}{Horizontal justification for title, subtitle, caption.}

\item{leg}{Legend position. "v" for regular vertical on the right. "h" for horizontal on top.}

\item{leg_just}{Legend justification}
}
\description{
ggplot2 custom theme using default font.
}
