% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_barlow.r, R/theme_inter.r,
%   R/theme_nunito.r, R/theme_plex.r
\name{theme_barlow}
\alias{theme_barlow}
\alias{theme_inter}
\alias{theme_nunito}
\alias{theme_plex}
\title{ggplot2 theme using custom font}
\usage{
theme_barlow(...)

theme_inter(...)

theme_nunito(...)

theme_plex(...)
}
\arguments{
\item{...}{Arguments passed to `plummy::theme_custom()`}
}
\description{
ggplot2 theme using custom font

ggplot2 theme using custom font

ggplot2 theme using custom font

ggplot2 theme using custom font
}
\note{
You should import fonts first and also install the fonts on your
system before trying to use this theme.
}
