% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_barlow.r, R/theme_grid.r,
%   R/theme_inter.r, R/theme_lato.r, R/theme_nunito.r, R/theme_plex.r
\name{theme_grid_barlow}
\alias{theme_grid_barlow}
\alias{theme_grid}
\alias{theme_grid_inter}
\alias{theme_grid_lato}
\alias{theme_grid_nunito}
\alias{theme_grid_plex}
\title{ggplot2 grid theme.}
\usage{
theme_grid_barlow(...)

theme_grid(
  base_family = "",
  base_size = 10.5,
  scaling = 1.2,
  title_family = "",
  title_size = base_size * scaling^2,
  title_face = "bold",
  title_color = "#2e4053",
  title_margin = margin(b = 2),
  subtitle_family = "",
  subtitle_size = base_size * scaling,
  subtitle_face = "plain",
  subtitle_color = "#2e4053",
  subtitle_margin = margin(b = 6),
  strip_text_family = "",
  strip_text_size = base_size,
  strip_text_face = "bold",
  strip_text_color = "#2e4053",
  caption_family = "",
  caption_size = base_size/scaling,
  caption_face = "plain",
  caption_color = "#aeb6bf",
  caption_margin = margin(t = 10),
  axis_text_family = "",
  axis_text_face = "plain",
  axis_text_color = "#2e4053",
  axis_text_size = base_size/scaling,
  axis_title_family = "",
  axis_title_size = base_size/scaling,
  axis_title_face = "plain",
  axis_title_color = "#2e4053",
  axis_title_just = "rt",
  plot_margin = margin(20, 20, 20, 20),
  plot_title_position = "plot",
  plot_caption_position = "plot",
  grid_color = "#cccccc",
  grid = TRUE,
  axis_color = "#cccccc",
  axis = TRUE,
  ticks = FALSE,
  tsc_just = 0,
  leg = "v",
  leg_just = "center",
  base_theme = theme_minimal
)

theme_grid_inter(...)

theme_grid_lato(...)

theme_grid_nunito(...)

theme_grid_plex(...)
}
\arguments{
\item{...}{Arguments passed to \code{plummy::theme_grid()}}

\item{base_family, base_size}{Base: font family and size}

\item{scaling}{Font size scaling factor}

\item{title_family, title_face, title_color, title_size, title_margin}{Plot title: font family, face, color, size and margin}

\item{subtitle_family, subtitle_face, subtitle_color, subtitle_size, subtitle_margin}{Plot subtitle: font family, face, color, size and margin}

\item{strip_text_family, strip_text_face, strip_text_color, strip_text_size}{Facet label: font family, face, color and size}

\item{caption_family, caption_face, caption_color, caption_size, caption_margin}{Plot caption: font family, face, color, size and margin}

\item{axis_text_family, axis_text_face, axis_text_color}{Axis text: font family, face and color}

\item{axis_text_size}{Font size of axis text}

\item{axis_title_family, axis_title_face, axis_title_color, axis_title_size}{Axis title: font family, face, color and size}

\item{axis_title_just}{Axis title font justification, one of \verb{[blmcrt]} denoting bottom, left, middle, center, right and top.}

\item{plot_margin}{Plot margin}

\item{plot_title_position}{Position of title/subtitle relative to panel (\code{"panel"}) or full plot (\code{"plot"}).}

\item{plot_caption_position}{Position of caption relative to panel (\code{"panel"}) or full plot (\code{"plot"}).}

\item{grid_color}{Grid color}

\item{grid}{Panel grid (\code{TRUE}, \code{FALSE}, or a combination of \code{X}, \code{x}, \code{Y}, \code{y} denoting major and minor gridlines)}

\item{axis_color}{Axis color}

\item{axis}{Add x or y axes? \code{TRUE}, \code{FALSE}, "\code{xy}". Includes axis title and axis text.}

\item{ticks}{Adds ticks if \code{TRUE}}

\item{tsc_just}{Horizontal justification for title, subtitle, caption.}

\item{leg}{Legend position. "v" for regular vertical on the right. "h" for horizontal on top.}

\item{leg_just}{Legend justification}

\item{base_theme}{Base theme. Defaults to \code{theme_minimal()}.}
}
\description{
ggplot2 grid theme.
}
\details{
Specify margins using \code{ggplot2::margin()}
}
