% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plummy.R
\name{update_geom_font}
\alias{update_geom_font}
\title{Update font for geom}
\usage{
update_geom_font(family = "")
}
\arguments{
\item{family}{A valid font family name}
}
\description{
Update font family for geom_text and geom_label
Note that this changes the default geom family for the whole session
}
