% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_barlow.r, R/theme_inter.r,
%   R/theme_nunito.r, R/theme_plex.r
\name{import_barlow}
\alias{import_barlow}
\alias{import_inter}
\alias{import_nunito}
\alias{import_plex}
\title{Import font for use in charts}
\usage{
import_barlow()

import_inter()

import_nunito()

import_plex()
}
\description{
Import font for use in charts

Import font for use in charts

Import font for use in charts

Import font for use in charts
}
\note{
Remember to run \code{extrafont::loadfonts()} after import. The location of the
font directory is displayed after the base import is complete. It is highly
recommended that you install them on your system the same way you would any
other font you wish to use in other programs.
}
