% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.r
\name{list_fonts}
\alias{list_fonts}
\title{List all fonts}
\usage{
list_fonts()
}
\value{
Returns a data.frame with font, family and variant. The font is used for plotting.
}
\description{
Lists all fonts, families and variants
}
